requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
charset-normalizer==3.2.0
click==8.1.4
exceptiongroup==1.1.2
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.7.1
yarl==1.9.2
